pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "binp51"
version = "0.1.0"
description = "eDNA biogeography"
readme = "README.md"
requires-python = ">=3.9"

# CORE: Only essential, conflict-free packages
dependencies = [
    "pandas>=1.3",
    "snakemake>=7.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.2",
]

[project.optional-dependencies]
# Development
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "ruff>=0.1.0",
    "black>=23.0",
    "isort>=5.0",
]

# Backend (FastAPI + core ML)
backend = [
    "fastapi>=0.131.0",
    "uvicorn>=0.41.0",
    "pydantic>=2.0",
]

# RAG system (docling + chromadb)
rag = [
    "chromadb>=1.5.2",
    "docling>=2.65",
    "openai>=2.0",
    "transformers>=4.42.0,<5.0",  
    "sentence-transformers>=5.1.0",
]

# DNABERTs (DNA embeddings)
dnaberts = [
    "torch>=2.0",
    "transformers>=5.3.0",
    "biopython>=1.86",
    "einops>=0.8.0",
]

# For local development: everything loose (pip resolves conflicts)
all = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "ruff>=0.1.0",
    "black>=23.0",
    "fastapi>=0.131.0",
    "uvicorn>=0.41.0",
    "chromadb>=1.5.2",
    "docling>=2.65",
    "openai>=2.0",
    "torch>=2.0",
    "transformers>=4.42.0",  
    "sentence-transformers>=5.1.0",
    "biopython>=1.86",
    "numpy<2.0,>=1.17",  
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 200
select = ["E", "F", "C", "B", "I"]
ignore = ["E203"]

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --tb=short"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
